+++ CHANGELOG.rst
@@ -6,6 +6,87 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog`_.
 
+1.2.0 (2022-02-10)
+==================
+
+Added
+-----
+
+- Kestrel main package
+
+  - matplotlib figure support in Kestrel Display Objects
+  - analytics interface upgraded with config shared to Kestrel
+    
+- Python analytics interface
+
+  - minimal requirement design for writing a Python analytics
+  - analytics function environment setup and destroy
+  - support for a variety of display object outputs
+  - parameters support
+  - stack tracing for exception inside a Python analytics
+    
+- STIX-shifter data source interface
+
+  - automatic STIX-shifter connector install
+    
+    - connector name guess
+    - connector origin verification
+    - comprehensive error and suggestion if automatic install failed
+        
+  - pretty print for exception inside a Docker analytics
+    
+- documentation
+
+  - Python analytics interface
+  - Kestrel debug page
+  - flag to disable certificate verification in STIX-shifter profile example
+
+Changed
+-------
+
+- abstract interface manager between datasource/analytics for code reuse
+
+Fixed
+-----
+
+- auto-complete with data source #163
+- exception for empty STIX-shifter profile
+- STIX-shifter profile name should be case insensitive
+- exception inappropriately caught when dereferencing vars with no time range
+
+Removed
+-------
+
+- documentation about STIX-shifter connector install
+
+1.1.7 (2022-01-27)
+==================
+
+Added
+-----
+
+- standalone Kestrel config module to support modular and simplified Kestrel config loading flow
+- shareable-state of config between Kestrel session and any Kestrel data source interfaces
+- stix-shifter interface upgraded with shareable-state of config support
+- stix-shifter DEBUG level env var ``KESTREL_STIXSHIFTER_DEBUG``
+- stix-shifter config/profile loading from disk ``~/.config/kestrel/stixshifter.yaml``
+- debug message logging in ``kestrel_datasource_stixshifter``
+- documentation for Kestrel main config with default config linked/shown
+
+Changed
+-------
+
+- default Kestrel config not managed by ``pip`` any more
+- turn main Kestrel from TOML into YAML ``~/.config/kestrel/kestrel.yaml``
+- upgrade Kestrel data source interfaces API with new ``config`` parameter
+- default stix-shifter debug level to INFO
+- documentation upgrade for ``kestrel_datasource_stixshifter``
+
+Fixed
+-----
+
+- Kestrel config upgrade inconsistency #116
+
 1.1.6 (2021-12-15)
 ==================